aliuacademy_org/academy/requirements/local.txt
-e .
factory-boy
pytest
pytest-cov
pytest-django
pytest-flakes
pytest-pep8
sphinx
